Upgrade to Visual Studio 2017
Note: zlib is compiled against the v141_winxp toolchain, because setting
the SDK only means you need that SDK installed or a higher version of
windows (Windows 8 in this case). Unfortunately setting the WINVER
macros doesn't seem to work when targetting the higher versions of
windows. (When setting them in property pages).